feat(dynamic-menu): moved from old to new MenuItem type
